+++ CPAC/pipeline/engine.py
@@ -1,41 +1,39 @@
-import os
 import ast
-import six
-import json
-import warnings
 import logging
+import os
+import warnings
 import copy
-from unittest import TestCase
+import yaml
 
-from CPAC.pipeline import nipype_pipeline_engine as pe
-import nipype.interfaces.utility as util
-from nipype.interfaces.utility import Rename
-from CPAC.utils.interfaces.function import Function
-from CPAC.utils.interfaces.datasink import DataSink
+from CPAC.pipeline import \
+    nipype_pipeline_engine as pe  # pylint: disable=ungrouped-imports
+from nipype.interfaces.utility import \
+    Rename  # pylint: disable=wrong-import-order
 
+from CPAC.image_utils.spatial_smoothing import spatial_smoothing
+from CPAC.image_utils.statistical_transforms import z_score_standardize, \
+    fisher_z_score_standardize
 from CPAC.registration.registration import transform_derivative
-from CPAC.nuisance import NuisanceRegressor
-
 from CPAC.utils import Outputs
-from CPAC.utils.utils import read_json, create_id_string, write_output_json, \
-    get_last_prov_entry, ordereddict_to_dict, check_prov_for_regtool
 from CPAC.utils.datasource import (
     create_anat_datasource,
     create_func_datasource,
     ingress_func_metadata,
     create_general_datasource,
-    create_check_for_s3_node,
     resolve_resolution
 )
-from CPAC.image_utils.spatial_smoothing import spatial_smoothing
-from CPAC.image_utils.statistical_transforms import z_score_standardize, \
-    fisher_z_score_standardize
+from CPAC.utils.interfaces.function import Function
+from CPAC.utils.interfaces.datasink import DataSink
+from CPAC.utils.monitoring.custom_logging import set_up_logger
+from CPAC.utils.utils import read_json, create_id_string, write_output_json, \
+    get_last_prov_entry, check_prov_for_regtool
 
 logger = logging.getLogger('nipype.workflow')
 verbose_logger = logging.getLogger('engine')
+outputs_logger = set_up_logger('expected_outputs')
 
 
-class ResourcePool(object):
+class ResourcePool:
     def __init__(self, rpool=None, name=None, cfg=None, pipe_list=None):
 
         if not rpool:
@@ -754,17 +752,56 @@ class ResourcePool(object):
         return wf
 
     def gather_pipes(self, wf, cfg, all=False, add_incl=None, add_excl=None):
-       
+        class ExpectedOutputs:
+            '''Class to hold expected outputs for a pipeline
+
+            Attributes
+            ----------
+            expected_outputs : dict
+                dictionary of expected output subdirectories and files therein
+
+            Methods
+            -------
+            add(subdir, output)
+                Add an expected output to the expected outputs dictionary
+            '''
+            def __init__(self):
+                self.expected_outputs = {}
+
+            def __repr__(self):
+                return self.__str__()
+
+            def __str__(self):
+                return yaml.dump(self.expected_outputs)
+
+            def add(self, subdir, output):
+                '''Add an expected output to the expected outputs dictionary
+
+                Parameters
+                ----------
+                subdir : str
+                    subdirectory of expected output
+
+                output : str
+                    filename of expected output
+                '''
+                if subdir in self.expected_outputs:
+                    self.expected_outputs[subdir].append(output)
+                else:
+                    self.expected_outputs[subdir] = [output]
+
         excl = []
         substring_excl = []
+        expected_outputs = ExpectedOutputs()
 
         if add_excl:
             excl += add_excl
-                       
-        if 'unsmoothed' not in cfg.post_processing['spatial_smoothing']['output']:
+
+        if 'unsmoothed' not in cfg.post_processing['spatial_smoothing'][
+                'output']:
             excl += Outputs.native_nonsmooth
             excl += Outputs.template_nonsmooth
-            
+
         if 'raw' not in cfg.post_processing['z-scoring']['output']:
             excl += Outputs.native_raw
             excl += Outputs.template_raw
@@ -897,6 +934,7 @@ class ResourcePool(object):
                                                             newdesc_suff)
                 else:
                     resource_idx = resource
+                expected_outputs.add(out_dct['subdir'], out_dct['filename'])
 
                 id_string = pe.Node(Function(input_names=['unique_id',
                                                           'resource',
@@ -992,6 +1030,8 @@ class ResourcePool(object):
                 wf.connect(write_json, 'json_file',
                            ds, f'{out_dct["subdir"]}.@json')
 
+        outputs_logger.info(expected_outputs)
+
     def node_data(self, resource, **kwargs):
         '''Factory function to create NodeData objects
 
@@ -1006,7 +1046,7 @@ class ResourcePool(object):
         return NodeData(self, resource, **kwargs)
 
 
-class NodeBlock(object):
+class NodeBlock:
     def __init__(self, node_block_functions):
 
         if not isinstance(node_block_functions, list):